Remove dest-file on fail & trimming url (close #203) (#204)
- trim version of node before build target download-url
- remove dest-file in case if download fails

+++ nj-core/src/basic.rs
@@ -633,7 +633,7 @@ impl JsEnv {
             &mut len
         ))?;
 
-        let array: &[u8] = unsafe { slice::from_raw_parts(data as *const u8, len as usize) };
+        let array: &[u8] = unsafe { slice::from_raw_parts(data as *const u8, len) };
 
         Ok(array)
     }
